+++ tests/test_unit_glms.py
@@ -18,23 +18,6 @@ from scipy.stats.distributions import poisson, nbinom, bernoulli, gamma
 
 np.random.seed(10)
 
-@pytest.fixture
-def scotland_data():
-    data = sm.datasets.scotland.load(as_pandas=False)
-    return data.exog, data.endog
-
-@pytest.fixture
-def poisson_data():
-    n_samples = 1000
-    int_coef, age_coef, weight_coef = -10, 0.05, 0.08
-    age = np.random.uniform(30, 70, n_samples)
-    weight = np.random.normal(150, 20, n_samples)
-    expected_visits = np.exp(int_coef + age * age_coef + weight * weight_coef)
-    observed_visits = poisson.rvs(expected_visits)
-    X = np.vstack([age, weight]).T
-    y = observed_visits
-    return X, y
-
 def test_gamma_example(scotland_data):
     """
     recreate the example here but through glimpy
@@ -61,45 +44,6 @@ def test_poisson_example(poisson_data):
     poisson_glm.fit(X, y)
     assert np.all(np.round(poisson_glm.coef_, 2) == [0.05, 0.08])
     assert round(poisson_glm.intercept_) == -10
-
-def test_l1_poisson(poisson_data):
-    """
-    test a poisson model
-    """
-    X, y = poisson_data
-    poisson_glm = GLM(fit_intercept=True, family=Poisson(),
-        penalty='l1', C=0.0001)
-    poisson_glm.fit(X, y)
-    assert poisson_glm.coef_[0] == 0
-    assert np.abs(poisson_glm.intercept_) > 1
-    assert poisson_glm.l1_ratio == 1
-
-def test_l2_poisson(poisson_data):
-    """
-    test a poisson model with l2 regularization
-    """
-    X, y = poisson_data
-    poisson_glm = GLM(fit_intercept=False, family=Poisson(),
-        penalty='l2', C=0.001)
-    intercept_col = np.ones(len(X)).reshape(-1, 1)
-    X = np.hstack([intercept_col, X])
-    poisson_glm.fit(X, y)
-    assert np.all(np.round(poisson_glm.coef_, 2) == [-0.03, 0.01, 0.03])
-    assert poisson_glm.l1_ratio == 0
-
-def test_elasticnet_poisson(poisson_data):
-    """
-    test a poisson model with l2 regularization
-    """
-    X, y = poisson_data
-    poisson_glm = GLM(fit_intercept=False, family=Poisson(),
-        penalty='elasticnet', C=0.001)
-    intercept_col = np.ones(len(X)).reshape(-1, 1)
-    X = np.hstack([intercept_col, X])
-    poisson_glm.fit(X, y)
-    assert np.all(np.round(poisson_glm.coef_, 2) == [0, 0, 0.04])
-    assert poisson_glm.l1_ratio == 0.5
-
 
 def test_irls_init():
     """
